Separated load trajectory node into a trajectory loading node and a node that extracts an approach process and departure trajectory from a nominal trajectory
