+++ snp_application/include/snp_application/bt/load_trajectory_from_file_node.h
@@ -4,13 +4,14 @@
 
 namespace snp_application
 {
+/**
+ * @brief Loads a trajectory from a YAML file in to a `joint_trajectory_msgs/JointTrajectory` message
+ */
 class LoadTrajectoryFromFileNode : public BT::SyncActionNode
 {
 public:
   inline static std::string FILE_NAME_INPUT_PORT_KEY = "file";
-  inline static std::string APPROACH_OUTPUT_PORT_KEY = "approach";
-  inline static std::string PROCESS_OUTPUT_PORT_KEY = "process";
-  inline static std::string DEPARTURE_OUTPUT_PORT_KEY = "departure";
+  inline static std::string TRAJECTORY_OUTPUT_PORT_KEY = "approach";
   static BT::PortsList providedPorts();
 
   explicit LoadTrajectoryFromFileNode(const std::string& instance_name, const BT::NodeConfig& config);
