+++ snp_application/src/bt/plugins.cpp
@@ -1,5 +1,6 @@
 #include <snp_application/bt/button_approval_node.h>
 #include <snp_application/bt/button_monitor_node.h>
+#include <snp_application/bt/extract_approach_process_departure_trajectories_node.h>
 #include <snp_application/bt/load_trajectory_from_file_node.h>
 #include <snp_application/bt/progress_decorator_node.h>
 #include <snp_application/bt/set_page_decorator_node.h>
@@ -20,6 +21,8 @@ BT_REGISTER_NODES(factory)
 {
   factory.registerNodeType<snp_application::ButtonApprovalNode>("ButtonApproval");
   factory.registerNodeType<snp_application::ButtonMonitorNode>("ButtonMonitor");
+  factory.registerNodeType<snp_application::ExtractApproachProcessDepartureTrajectoriesNode>(
+      "ExtractApproachProcessDepartureTrajectories");
   factory.registerNodeType<snp_application::LoadTrajectoryFromFileNode>("LoadTrajectoryFromFile");
   factory.registerNodeType<snp_application::ProgressDecoratorNode>("Progress");
   factory.registerNodeType<snp_application::SetPageDecoratorNode>("SetPage");
